User Feedback: Natural Percent Order

+++ iBroker/Calculator/Calculator.h
@@ -16,9 +16,9 @@
  *
  * @param row
  * @param index
- * @return
+ * @return NSString*
  */
-+ (NSString *)assetString:(long)row withIndex:(long)index;
++ (NSString *)assetString:(unsigned int)row withIndex:(unsigned int)index;
 
 /**
  * Static Constructor implemented as singleton
